Added a new job before fetching seeds to clearcut the zone.
Behavior is when a pawn considers fetching a seed to start sowing they first check the zone for any plants that don't belong as well as anything adjacent to the zone which would block sowing.  If something is found they will cut it all before fetching the seed.

Once a pawn has a seed in hand they will keep sowing until they run out of spaces or seeds at which time (if they want to keep sowing) they will do the check for if there are plants that don't belong.